pyproject.toml
[project]
name = "hindi-srs"
version = "0.1.0"
description = "Spaced repetition language learning system for Hindi, powered by LLM-generated exercises"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.14.0",
    "anthropic>=0.42.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "tenacity>=9.0.0",
    "aiosqlite>=0.20.0",
    "Pillow>=10.0.0",
    "greenlet>=3.3.1",
]

[project.optional-dependencies]
pdf = [
    "pdfplumber>=0.11.0",
]
dev = [
    "ruff>=0.9.0",
    "mypy>=1.14.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=4.0.0",
    "httpx>=0.28.0",
    "pre-commit>=4.0.0",
    "pdfplumber>=0.11.0",
    "fpdf2>=2.8.0",
    "bandit>=1.7.0",
]

[project.scripts]
hindi-srs = "hindi_srs.__main__:main"
hindi-srs-ingest = "scripts.ingest:main"
hindi-srs-load = "scripts.load_to_db:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "W",    # pycodestyle warnings
    "I",    # isort
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "D",    # pydocstyle (docstrings)
    "C90",  # mccabe complexity
    "PTH",  # flake8-use-pathlib
    "RUF",  # ruff-specific rules
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["D"]  # No docstrings required in tests
"scripts/*" = ["D"]  # No docstrings required in scripts
"migrations/*" = ["D"]  # Alembic auto-generated files

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.bandit]
exclude_dirs = ["tests", "migrations"]
skips = ["B101"]  # Allow assert statements
